Microserviciu3/app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
httpx
grpcio
grpcio-tools
pyjwt
python-multipart
bcrypt
protobuf